[Refactor] Analytics 모듈 domain/data 분리
