+++ app/build.gradle.kts
@@ -67,7 +67,7 @@ dependencies {
     implementation(projects.core.data)
     implementation(projects.core.domain)
     implementation(projects.core.alarm)
-    implementation(projects.core.analytics)
+    implementation(projects.core.analytics.data)
     implementation(projects.core.experiment.data)
 
     implementation(projects.feature.onboarding)
