+++ core/analytics/data/src/main/java/com/tgyuu/analytics/data/AmplitudeAnalyticsHelper.kt
@@ -1,12 +1,14 @@
-package com.tgyuu.analytics
+package com.tgyuu.analytics.data
 
 import com.amplitude.android.Amplitude
 import com.amplitude.core.events.BaseEvent
+import com.tgyuu.analytics.domain.AnalyticsHelper
+import com.tgyuu.analytics.domain.model.AnalyticsEvent
 import javax.inject.Inject
 
 class AmplitudeAnalyticsHelper @Inject constructor(
     private val amplitude: Amplitude,
-) : AnalyticsHelper() {
+) : AnalyticsHelper {
     override fun logEvent(event: AnalyticsEvent) {
         amplitude.track(event = event.toAmplitudeEvent())
     }
