+++ core/analytics/data/src/main/java/com/tgyuu/analytics/data/NoOpAnalyticsHelper.kt
@@ -1,0 +1,9 @@
+package com.tgyuu.analytics.data
+
+import com.tgyuu.analytics.domain.AnalyticsHelper
+import com.tgyuu.analytics.domain.model.AnalyticsEvent
+
+class NoOpAnalyticsHelper : AnalyticsHelper {
+    override fun logEvent(event: AnalyticsEvent) = Unit
+    override fun setUserId(userId: String?) = Unit
+}
