+++ core/analytics/domain/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("ebbing.kotlin.library")
+}
