+++ core/analytics/domain/src/main/java/com/tgyuu/analytics/domain/AnalyticsHelper.kt
@@ -1,0 +1,8 @@
+package com.tgyuu.analytics.domain
+
+import com.tgyuu.analytics.domain.model.AnalyticsEvent
+
+interface AnalyticsHelper {
+    fun logEvent(event: AnalyticsEvent)
+    fun setUserId(userId: String?)
+}
