+++ core/analytics/domain/src/main/java/com/tgyuu/analytics/domain/model/AnalyticsEvent.kt
@@ -1,4 +1,4 @@
-package com.tgyuu.analytics
+package com.tgyuu.analytics.domain.model
 
 data class AnalyticsEvent(
     val type: String,
