+++ feature/home/build.gradle.kts
@@ -20,7 +20,7 @@ android {
 dependencies {
     implementation(projects.core.common)
     implementation(projects.core.alarm)
-    implementation(projects.core.analytics)
+    implementation(projects.core.analytics.data)
     implementation(projects.core.experiment.domain)
 
     implementation(libs.accompanist.permission)
