+++ settings.gradle.kts
@@ -34,7 +34,8 @@ include(":core:navigation")
 include(":core:database")
 include(":core:datastore")
 include(":core:alarm")
-include(":core:analytics")
+include(":core:analytics:domain")
+include(":core:analytics:data")
 include(":core:experiment:domain")
 include(":core:experiment:data")
 
@@ -48,4 +49,3 @@ include(":feature:repeatcycle")
 include(":feature:sync")
 
 include(":baselineprofile")
-include(":core:analytics")
